+++ data/src/main/java/com/anytypeio/anytype/data/auth/repo/block/BlockDataStore.kt
@@ -141,7 +141,8 @@ interface BlockDataStore {
         filters: List<DVFilter>,
         fulltext: String,
         offset: Int,
-        limit: Int
+        limit: Int,
+        keys: List<Id>
     ): List<Map<String, Any?>>
 
     suspend fun searchObjectsWithSubscription(