+++ data/src/main/java/com/anytypeio/anytype/data/auth/repo/block/BlockRemote.kt
@@ -147,7 +147,8 @@ interface BlockRemote {
         filters: List<DVFilter>,
         fulltext: String,
         offset: Int,
-        limit: Int
+        limit: Int,
+        keys: List<Id>
     ): List<Map<String, Any?>>
 
     suspend fun searchObjectsWithSubscription(