+++ middleware/src/main/java/com/anytypeio/anytype/middleware/block/BlockMiddleware.kt
@@ -355,13 +355,15 @@ class BlockMiddleware(
         filters: List<DVFilter>,
         fulltext: String,
         offset: Int,
-        limit: Int
+        limit: Int,
+        keys: List<Id>
     ): List<Map<String, Any?>> = middleware.searchObjects(
         sorts = sorts,
         filters = filters,
         fulltext = fulltext,
         offset = offset,
-        limit = limit
+        limit = limit,
+        keys = keys
     )
 
     override suspend fun searchObjectsWithSubscription(