+++ presentation/src/main/java/com/anytypeio/anytype/presentation/search/ObjectSearchViewModel.kt
@@ -73,7 +73,8 @@ open class ObjectSearchViewModel(
         limit = SEARCH_LIMIT,
         filters = ObjectSearchConstants.filterSearchObjects,
         sorts = ObjectSearchConstants.sortsSearchObjects,
-        fulltext = EMPTY_QUERY
+        fulltext = EMPTY_QUERY,
+        keys = ObjectSearchConstants.defaultKeys
     )
 
     private fun startProcessingSearchQuery() {